+++ src/rwa/components/group-transaction-details/group-transaction-details.tsx
