+++ src/rwa/components/group-transaction-details/index.ts
@@ -1,0 +1,1 @@
+export * from './group-transaction-details';
