+++ src/rwa/components/index.ts
@@ -1,10 +1,10 @@
 export * from './asset-details';
 export * from './button';
 export * from './date-picker';
+export * from './group-transaction-details';
 export * from './select';
 export * from './table';
 export * from './tabs';
 export * from './text-input';
-export * from './tx-detail';
 
 export * from './button';
