--- src/rwa/components/tx-detail/form.ts
@@ -1,4 +1,0 @@
-export type RWATxSelectTypes = 'assetTypeId' | 'cusipIsinAssetNameId';
-export type RWATxTextTypes = 'transactionType' | 'assetProceedsUSD';
-export type RWATxDateTypes = 'timestamp';
-export type RWATxFeesTypes = 'feesTable';
