--- src/rwa/components/tx-detail/index.ts
@@ -1,1 +1,0 @@
-export * from './tx-detail';
